Trim the tool-registry import chain of nio and agno runtime imports
Importing tool_system.metadata pulled in the nio matrix client (via
runtime_context -> attachments, hooks.context -> matrix_admin,
hooks.state, and message_target -> thread_utils -> authorization) and
the agno tools runtime (via output_files). Move the tiny attachment-ID
and session-ID helpers into leaf modules, defer the nio-touching hook
imports to first use, and import output_files at first tool
construction, so the registry (metadata + catalog) imports stay slim
for template warm-up, pod boot, and CI collection.

Part of #1436.
