+++ src/mindroom/api/sandbox_runner.py
@@ -28,7 +28,7 @@ from mindroom.api.worker_responses import (
     SandboxWorkerListResponse,
     serialize_sandbox_worker_response,
 )
-from mindroom.attachments import normalize_attachment_id
+from mindroom.attachment_ids import normalize_attachment_id
 from mindroom.config.main import Config, load_config, normalized_config_data
 from mindroom.config.yaml_includes import load_yaml_config_source
 from mindroom.credentials import CredentialsManager, get_runtime_credentials_manager, load_scoped_credentials
