+++ src/mindroom/attachments.py
@@ -7,7 +7,6 @@ import contextlib
 import hashlib
 import json
 import mimetypes
-import re
 import time
 from dataclasses import dataclass
 from datetime import UTC, datetime, timedelta
@@ -17,6 +16,7 @@ from uuid import uuid4
 
 import nio
 
+from .attachment_ids import normalize_attachment_id
 from .constants import ATTACHMENT_IDS_KEY
 from .logging_config import get_logger
 from .matrix.media import (
@@ -37,25 +37,16 @@ from .matrix.media import (
 from .timing import emit_elapsed_timing
 
 if TYPE_CHECKING:
-    from collections.abc import Iterable, Sequence
+    from collections.abc import Sequence
 
     from .matrix.client import ResolvedVisibleMessage
 
 logger = get_logger(__name__)
 
 _AttachmentKind = Literal["audio", "file", "image", "video"]
-_ATTACHMENT_ID_PATTERN = re.compile(r"^[A-Za-z0-9_][A-Za-z0-9_-]{0,127}$")
 _ATTACHMENT_RETENTION_DAYS = 30
 _CLEANUP_INTERVAL = timedelta(hours=1)
 _last_cleanup_time_by_storage_path: dict[Path, datetime] = {}
-
-
-def normalize_attachment_id(raw_attachment_id: str) -> str | None:
-    """Normalize attachment IDs and reject unsafe values."""
-    attachment_id = raw_attachment_id.strip()
-    if not attachment_id or not _ATTACHMENT_ID_PATTERN.fullmatch(attachment_id):
-        return None
-    return attachment_id
 
 
 @dataclass(frozen=True)
@@ -180,24 +171,6 @@ def attachment_records_for_visible_message(
         return []
     records = resolve_attachments(storage_path, attachment_ids)
     return [record for record in records if _attachment_record_in_message_scope(record, message, room_id=room_id)]
-
-
-def unique_attachment_ids(attachment_ids: Iterable[str]) -> list[str]:
-    """Return unique non-empty attachment IDs preserving first-seen order."""
-    unique_ids: list[str] = []
-    seen_attachment_ids: set[str] = set()
-    for attachment_id in attachment_ids:
-        if attachment_id and attachment_id not in seen_attachment_ids:
-            seen_attachment_ids.add(attachment_id)
-            unique_ids.append(attachment_id)
-    return unique_ids
-
-
-def merge_attachment_ids(*attachment_id_lists: list[str]) -> list[str]:
-    """Merge attachment IDs preserving first-seen order."""
-    return unique_attachment_ids(
-        attachment_id for attachment_ids in attachment_id_lists for attachment_id in attachment_ids
-    )
 
 
 _MAX_RENDERED_FILENAME_LENGTH = 80
