+++ src/mindroom/hooks/context.py
@@ -11,7 +11,6 @@ from mindroom.logging_config import get_logger
 from mindroom.runtime_protocols import SupportsClientConfigOrchestrator  # noqa: TC001
 from mindroom.tool_system.plugin_identity import validate_plugin_name
 
-from . import matrix_admin as hook_matrix_admin
 from .state import (
     build_hook_room_state_putter,
     build_hook_room_state_querier,
@@ -197,7 +196,11 @@ class HookContextSupport:
             if admin is not None:
                 return admin
         if self.agent_name == ROUTER_AGENT_NAME and self.runtime.client is not None:
-            return hook_matrix_admin.build_hook_matrix_admin(
+            # Imported on first use so the tool-registry import chain stays free
+            # of the nio matrix-client import (#1436).
+            from .matrix_admin import build_hook_matrix_admin  # noqa: PLC0415
+
+            return build_hook_matrix_admin(
                 self.runtime.client,
                 self.runtime_paths,
                 config=self.runtime.config,
