+++ src/mindroom/thread_utils.py
@@ -120,18 +120,6 @@ def check_agent_mentioned(
     return mentioned_agents, am_i_mentioned, has_non_agent_mentions
 
 
-def create_session_id(room_id: str, thread_id: str | None) -> str:
-    """Create a session ID with thread awareness."""
-    # Thread sessions include thread ID
-    return f"{room_id}:{thread_id}" if thread_id else room_id
-
-
-def parse_session_id(session_id: str) -> tuple[str, str | None]:
-    """Parse the canonical persisted room/thread session ID."""
-    room_id, marker, thread_suffix = session_id.rpartition(":$")
-    return (room_id, f"${thread_suffix}") if marker else (session_id, None)
-
-
 def get_agents_in_thread(
     thread_history: Sequence[ResolvedVisibleMessage],
     config: Config,
