+++ src/mindroom/turn_store.py
@@ -14,7 +14,7 @@ from mindroom import constants
 from mindroom.agent_storage import get_agent_session, get_team_session
 from mindroom.agents import remove_run_by_event_id
 from mindroom.handled_turns import HandledTurnLedger, HandledTurnRecord, HandledTurnState, SourceEventMetadata
-from mindroom.thread_utils import create_session_id
+from mindroom.session_ids import create_session_id
 
 if TYPE_CHECKING:
     import nio
